Refactor encryption utility out of shared utils package
Moved encryption-related functions from the shared `utils` package into the `encryptor` package to improve cohesion. Updated references in `service_impl.go` accordingly to use the local package functions.
